backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
numpy==2.2.6
opencv-python==4.12.0.88
requests==2.32.5
google-genai==1.32.0
